Sanctum of Rebirth
added Sanctum of Rebirth bosses, items, and setup items

+++ Assets/Scripts/BossInfo/Runtime/BossCombatData.cs
@@ -50,7 +50,7 @@ public class BossCombatData : IComparable<BossCombatData>
     /// <returns></returns>
     public double HitChance(AttackingPlayer player)
     {
-        return AffinityData.GetAffinity(player.Weapon, player.AffinityModifier) * (double)player.Accuracy() / ArmourRating();
+        return AffinityData.GetAffinity(player.Weapon, player.AffinityModifier) * player.Accuracy() / ArmourRating();
     }
 
     /// <summary>
